refactor!: Reduce work for Groth16 REP3 by working over un-replicated shares as much as possible (#349)
Since replicated secret sharing is just parties keeping additional shares around to facilitate multiplication in a more efficient way, we can not use this replication if we do not need any multiplications. In particular, for MSMs, it is much more efficient to reduce the replication down to 1 before doing the MSM, as the result is a single point share, which is very efficiently replicated again. This reduces the work done in MSMs in groth16 for rep3 sharing by a factor of about 2, which is a significant part of the overall runtime.

BREAKING CHANGE: Public API of coGroth16 driver trait changed to include `HalfShare` types.
---------

Co-authored-by: Franco Nieddu <nieddu@taceo.io>